+++ docs/changes.rst
@@ -20,6 +20,7 @@ v1.1.6 (in development)
 - New: ``oatmeal``
 - New: ``zelda``
 - Update: ``abstrusegoose`` has a schedule.
+- Update: ``apokalips`` is no longer published.
 - Update: ``asofterworld`` after feed change.
 - Update: ``axecop`` has a schedule.
 - Update: ``basicinstructions`` has a new schedule.